Tag amsmath equation environments
This adds an `currentequation` property to the Context object which stores
the last seen equation environment.

The `\tag` command sets the `equation_tag` property of the current
equation to its argument.

When getting the source of a MathEnvironment, a \tag command is
included, using either the `equation_tag`, or the element's `ref`
property, which will correspond to the numbering from the `equation`
counter, or `parentequation` and `equation` in a subequations
environment.

This doesn't deal with nested subequations environments. I was unsure
how to redefine the `theparentequation` command to do this, but hopefully
someone who knows the internals of plasTeX better will be able to sort
it out.
I think that `\theparentequation` should expand to whatever
`theequation` did at the start of the subequations environment, instead
of using the default definition which looks at the `parentequation`
counter.