Add support for custom item labels to enumerate
This adds support for custom item labels to enumerate environments,
following the format used by the enumerate LaTeX package.

The enumerate environment takes a single option argument specifying the
format for item labels. The format string can contain any characters;
the special characters I, i, A, a and 1 are replaced with corresponding
encodings of the item's position in the list.