beamer subclasses enumerate_ instead of modifying it
The unit test for the enumerate environment was failing when run after
the unit test for the beamer package, because beamer was modifying the
list of arguments for the `enumerate_` class.

This changes the beamer package to subclass `enumerate_`, instead of
just changing its args property.

The beamer package modifies a lot of other environments' arguments, so
this will affect any other unit tests written for them in the future.