(sub-sub-)paragraphs don't group their contents into paragraphs
In pdflatex, the \paragraph and \subparagraph commands work on the
current paragraph, and any inline text following them belongs to the
same paragraph.

This changes the paragraph, subparagraph and subsubparagraph commands so
that they don't group their contents into paragraphs, and are rendered
with a single <p> element starting with a <strong> element containing
the title.

Because HTML can't have heading elements inside a <p>, this means that
they're no longer in the heading hierarchy, but this rendering is closer
in other respects to how pdflatex renders them.