MathSymbols can be redefined.
This adds a flag to MathSymbol commands to say that they can be
redefined by LaTeX commands such as `\renewcommand`.

Context.newcommand looks for an attribute `redefinable` on the command
which is about to be redefined. This means that commands defined by
plasTeX can opt in to being redefined by the user.

This quickly gets round the problem I'm facing, but it might be that a
different model would work better for other similar situations.

See plastex/plastex#90.