strip trailing whitespace from the command name in \[re]newcommand
By default, plasTeX adds a space character to the end of the source of
any element that doesn't have any arguments, to ensure that it's
separated from the next token.

Most of the time this is safe because LaTeX usually ignores extra
whitespace. But the `\renewcommand` command is sensitive to extra
whitespace in the name of command being defined, when it's not a letter
character, e.g. `\renewcommand{\> }` is not interpreted the same as
`\renewcommand{\>}`.

I think that all of the `.source` methods on elements assume that they
don't need to know their outer context, so it was hard to think of a
good way of fixing this.

This commit adds a method `filterArgumentSource` to `Macro`, which
is called on the source of each argument as it's processed.

The `\newcommand` macro and all the others that inherit from it use this
to remove any whitespace from the inside of the first argument.